+++ cmd/frontend/internal/graphqlbackend/search_alert.go
@@ -329,13 +329,17 @@ outer:
 func (r *searchResolver) alertForMissingRepoRevs(missingRepoRevs []*repositoryRevisions) *searchAlert {
 	var description string
 	if len(missingRepoRevs) == 1 {
-		description = fmt.Sprintf("The repository %s matched by your repo: filter could not be searched because it does not contain the revision %q.", missingRepoRevs[0].repo.URI, missingRepoRevs[0].revspecs()[0])
+		if len(missingRepoRevs[0].revspecs()) == 1 {
+			description = fmt.Sprintf("The repository %s matched by your repo: filter could not be searched because it does not contain the revision %q.", missingRepoRevs[0].repo.URI, missingRepoRevs[0].revspecs()[0])
+		} else {
+			description = fmt.Sprintf("The repository %s matched by your repo: filter could not be searched because it has multiple specified revisions: @%s.", missingRepoRevs[0].repo.URI, strings.Join(missingRepoRevs[0].revspecs(), ","))
+		}
 	} else {
 		repoRevs := make([]string, 0, len(missingRepoRevs))
 		for _, r := range missingRepoRevs {
 			repoRevs = append(repoRevs, string(r.repo.URI)+"@"+strings.Join(r.revspecs(), ","))
 		}
-		description = fmt.Sprintf("%d repositories matched by your repo: filter could not be searched because the following revisions do not exist: %s.", len(missingRepoRevs), strings.Join(repoRevs, ", "))
+		description = fmt.Sprintf("%d repositories matched by your repo: filter could not be searched because the following revisions do not exist, or differ but were specified for the same repository: %s.", len(missingRepoRevs), strings.Join(repoRevs, ", "))
 	}
 	return &searchAlert{
 		title:       "Some repositories could not be searched",